src/loadgen/requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.6.3
certifi==2018.10.15
chardet==3.0.4
click==7.0
flask==1.0.2
gevent==1.3.7
greenlet==0.4.15 ; platform_python_implementation == 'CPython'
idna==2.7
itsdangerous==1.1.0
jinja2==2.10
locustio==0.9.0
markupsafe==1.1.0
msgpack==0.5.6
pyzmq==17.1.2
requests==2.20.1
six==1.11.0
urllib3==1.24.1
werkzeug==0.14.1
